+++ packages/extras/src/lib/components/InstancedSprite/SpriteInstance.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
   import { getContext } from 'svelte'
-  import type { InstancedSpriteUserCtx } from './InstancedSprite.svelte'
-  import type { SpriteInstanceProps } from './SpriteInstance.svelte'
+  import type { InstancedSpriteUserCtx, SpriteInstanceProps } from './types'
 
   let {
     id = 0,
